--- a/src/NewId/NewIdFormatters/DashedHexFormatter.cs
+++ b/src/NewId/NewIdFormatters/DashedHexFormatter.cs
@@ -1,12 +1,22 @@
 ﻿namespace MassTransit.NewIdFormatters
 {
+    using System;
+#if NET6_0_OR_GREATER
+    using System.Runtime.InteropServices;
+    using System.Runtime.CompilerServices;
+    using System.Runtime.Intrinsics;
+    using System.Runtime.Intrinsics.X86;
+#endif
+
+
     public class DashedHexFormatter :
         INewIdFormatter
     {
-        readonly int _alpha;
+        readonly uint _alpha;
         readonly int _length;
         readonly char _prefix;
         readonly char _suffix;
+        const uint LowerCaseUInt = 0x2020U;
 
         public DashedHexFormatter(char prefix = '\0', char suffix = '\0', bool upperCase = false)
         {
@@ -19,12 +29,22 @@
                 _length = 38;
             }
 
-            _alpha = upperCase ? 'A' : 'a';
+            _alpha = upperCase ? 0 : LowerCaseUInt;
         }
 
-        public string Format(in byte[] bytes)
+        public unsafe string Format(in byte[] bytes)
         {
-            var result = new char[_length];
+#if NET6_0_OR_GREATER
+            if (Avx2.IsSupported && BitConverter.IsLittleEndian)
+            {
+                var isUpperCase = _alpha != LowerCaseUInt;
+                return string.Create(_length, (bytes, isUpperCase, _prefix, _suffix), (span, state) =>
+                {
+                    EncodeVector256(span, state);
+                });
+            }
+#endif
+            var result = stackalloc char[_length];
 
             var i = 0;
             var offset = 0;
@@ -32,41 +52,41 @@
                 result[offset++] = _prefix;
             for (; i < 4; i++)
             {
-                int value = bytes[i];
-                result[offset++] = HexToChar(value >> 4, _alpha);
-                result[offset++] = HexToChar(value, _alpha);
+                var value = bytes[i];
+                HexToChar(value, result, offset, _alpha);
+                offset+=2;
             }
 
             result[offset++] = '-';
             for (; i < 6; i++)
             {
-                int value = bytes[i];
-                result[offset++] = HexToChar(value >> 4, _alpha);
-                result[offset++] = HexToChar(value, _alpha);
+                var value = bytes[i];
+                HexToChar(value, result, offset, _alpha);
+                offset+=2;
             }
 
             result[offset++] = '-';
             for (; i < 8; i++)
             {
-                int value = bytes[i];
-                result[offset++] = HexToChar(value >> 4, _alpha);
-                result[offset++] = HexToChar(value, _alpha);
+                var value = bytes[i];
+                HexToChar(value, result, offset, _alpha);
+                offset+=2;
             }
 
             result[offset++] = '-';
             for (; i < 10; i++)
             {
-                int value = bytes[i];
-                result[offset++] = HexToChar(value >> 4, _alpha);
-                result[offset++] = HexToChar(value, _alpha);
+                var value = bytes[i];
+                HexToChar(value, result, offset, _alpha);
+                offset+=2;
             }
 
             result[offset++] = '-';
             for (; i < 16; i++)
             {
-                int value = bytes[i];
-                result[offset++] = HexToChar(value >> 4, _alpha);
-                result[offset++] = HexToChar(value, _alpha);
+                var value = bytes[i];
+                HexToChar(value, result, offset, _alpha);
+                offset+=2;
             }
 
             if (_suffix != '\0')
@@ -75,10 +95,57 @@
             return new string(result, 0, _length);
         }
 
-        static char HexToChar(int value, int alpha)
+#if NET6_0_OR_GREATER
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static void EncodeVector256(Span<char> span, (byte[] bytes, bool, char _prefix, char _suffix) state)
         {
-            value &= 0xf;
-            return (char)(value > 9 ? value - 10 + alpha : value + 0x30);
+            var (bytes, isUpper, prefix, suffix) = state;
+            var swizzle = Vector256.Create((byte)
+                0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
+                0x80, 0x08, 0x09, 0x0a, 0x0b, 0x80, 0x0c, 0x0d,
+                0x80, 0x80, 0x80, 0x00, 0x01, 0x02, 0x03, 0x80,
+                0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b);
+
+            var dash = Vector256.Create((byte)
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x2d, 0x00, 0x00, 0x00, 0x00, 0x2d, 0x00, 0x00,
+                0x00, 0x00, 0x2d, 0x00, 0x00, 0x00, 0x00, 0x2d,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00);
+
+            var inputVec = MemoryMarshal.Read<Vector128<byte>>(bytes);
+            var hexVec = IntrinsicsHelper.EncodeBytesHex(inputVec, isUpper);
+
+            var a1 = Avx2.Shuffle(hexVec, swizzle);
+            var a2 = Avx2.Or(a1, dash);
+
+            if (span.Length == 38)
+            {
+                span[0] = prefix;
+                span[^1] = suffix;
+            }
+
+            var charSpan = span.Length == 38 ? span[1..^1] : span;
+            var spanBytes = MemoryMarshal.Cast<char, byte>(charSpan);
+            IntrinsicsHelper.Vector256ToCharUtf16(a2, spanBytes);
+
+            spanBytes[32] = hexVec.GetElement(14);
+            spanBytes[34] = hexVec.GetElement(15);
+
+            spanBytes[64] = hexVec.GetElement(28);
+            spanBytes[66] = hexVec.GetElement(29);
+            spanBytes[68] = hexVec.GetElement(30);
+            spanBytes[70] = hexVec.GetElement(31);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+#endif
+        static unsafe void HexToChar(byte value, char* buffer, int startingIndex, uint casing)
+        {
+            uint difference = (((uint)value & 0xF0U) << 4) + ((uint)value & 0x0FU) - 0x8989U;
+            uint packedResult = ((((uint)(-(int)difference) & 0x7070U) >> 4) + difference + 0xB9B9U) | (uint)casing;
+
+            buffer[startingIndex + 1] = (char)(packedResult & 0xFF);
+            buffer[startingIndex] = (char)(packedResult >> 8);
         }
     }
 }