… src/NewId/NewIdFormatters/ZBase32Formatter.cs
… tests/NewId.Tests/Formatter_Specs.cs
… tests/NewId.Tests/guids.txt
… tests/NewId.Tests/texts.txt