@phatboyg 2023-01-07
Quick question, with the changes to the Guids files that were used for testing, it seems like there is enough backwards compatible tests since those files were generated years ago. Only concern was compatibility with existing Guids, etc. written out as base32 but from your comment above, the mistake in the first byte is consistent and therefore kept in place. Solid effort, thanks!