@TimothyMakkison 2023-01-08
Yeah I though the encoding tests were a little overkill 😄. I generated the values on from an unmodified branch and used them to catch any errors. They could be removed/reduced if you think they cause bloat.

Note the `Base32Parser` could be vectorized but it doesn't appear to be used in `MassTransit`. It looks like the `ZBase32Parser` is used once in `FutureLocation`, I suspect its not possible to vectorize ZBase32 decoding so unless it's used much it's probably not worth it.