@phatboyg 2023-01-08
I appreciate the detail and comparison to values generated and formatted by the older v3 branch of NewId. Consistency is really important since changes could cause existing stored values to be lost, etc. 

Thank you again!